Modify zero-results m/r test to account for new not_found result
Riak now returns a not_found value for each {b, k} input to a m/r that does
not exist. The PB interface has a bug encoding this result. While
waiting for the fix for the PB interface, modify the zero-results test
to side-step the bug. Add a new test for not_found when the bug is fixed.