Make raw HTTP client expose riak config
Clients that sit on top of the Riak Java client (such as [Welle](http://clojureriak.info)) and provide
additional features, for example, Solr API support, may need to access HTTP endpoint.

In Welle, to support Solr API we currently have to subclass the raw HTTP client and use our own
class that does nothing else but exposes the config.

It would be nice if the Java client just provided access to it.