Fixed M/R logical keyfilters
This fixes a bug that causes the output JSON to look like:
["and",["starts_with","p"],["ends_with","1"]]
when it should look like:
["and",[["starts_with","p"]],[["ends_with","1"]]]

This was occuring with all logical key filters (and,or, and not)
and actually would cause the M/R job to crash in riak and timeout
on the client side.

The test was also modified to test for the correct structure.

Fixes #181