Change 2i to use long values
The original implementation of 2i used int for _int indexes
which is caused an upper bound of 2^31. The client will
now use long values eliminating this restriction.

When possible backwards compatability was maintained, specifically
regarding the @RiakIndex notation and the FetchIndex class. Users of
the IRiakObject interface calling getIntIndex() or allIntIndexes()
will need to change the type being returned from Set<Integer> to Set<Long>

This addresses #112