Added missing conditional
The logic here was meant to have the PB client behavior match the HTTP
client behavior when only fetching the meta and siblings were found to
be present. The conditional for seeing if the 'head' parameter had been
set was missing, causing every fetch that had siblings to be repeated.
This actually isn't necessarity with PB, but I don't want to change this
behavior now in case someone is relying on it.