1.4 feature adds streaming list buckets
This adds the 1.4 feature of streaming the bucket list.

It also fixes a long standing problem with streaming keys  in the
1.x client where there was no way to abort the operation without iterating
through the entire set, causing connections to get tied up. Both listing
keys and buckets now allows this by returning the new StreamingOperation
object which has a cancel() method.

This does change the RawClient interface. Projects built on top of
this interface will need to be adjusted accordingly.

Using the new 1.4 features against a < 1.4 Riak will throw exceptions.