Combined ConnectionPool and RiakNode
The two classes were inherently coupled after removing simultaneous,
multiple protocol support; You can't have a RiakNode without
a ConnectionPool and the ConnectionPool isn't used outside of a RiakNode.

In addition, having them as two classes created an abstraction layer that
added complexity and made a couple things clunky - notablly registering the RiakNode as a listener
to a specific channel handler in the pipeline.

I moved and modified all the ConnectionPool tests to operate on the RiakNode.