Fixed tombstone handling in DomainBucket
DomainBucket.fetch() was not honoring the ReturnDeletedVClock
setting in the FetchMeta.

Furter investigation found that there was no way to perform a delete
via DomainBucket using a vclock and a bug set the option for the Delete to
fetch the vclock to false.

DomainBucket.delete() now will either take a vclock or tell the
delete operation to fetch it first if it's not present.