TODO cleanup, slight healthcheck refactor
Removed a few TODO items that I feel are no longer issues.

Refactored healthcheck slightly. It now purges the recentlyClosed
sliding window first then makes a decision on whether to healthcheck
the node. Also decided that the timing and window are fine and
really don't need to be user settable (avoid ETTOMANYOPTIONS). We
know Riak's and the protocol's expected behavior.