Fixing broken itests.
Noticed a couple itests were failing. In fixing them I may have
uncovered a bug in the counter crdt. If you run ITestDtUpdateOperation
the testCrdtCounter() will fail. This is due to the fetch being
performed by L109 comes back with the PB message saying there is
a context (hasContext() == true) but the context is a zero-length
ByteString.