Move Location to core
Location is actually a very useful abstraction that I realized for
various reasons should be in core and then exposed at the user API
level.

This pushes it down into com.basho.client.core.query, makes it bit more
user friendly with String methods, and updates everything above to use it
unstead of the original.

In addition, StoreOperation and FetchOperation are updated accept it.