Moves conversion and resolution out to userland
This solves a number of ugly API issues the old client had.

The user API FetchValue, StoreValue, and UpdateValue are no longer generic
classes. The conflict resolution and conversion now happens when the user actually
gets the response values and they simply supply a class they'd like the response converted
to.

This is still a WIP as I need to add more tests.