Reworked 2i query at the User API level.
The FetchIndex code was broken in a couple places, and once I started
trying to fix it I just felt maybe there was a different approach.

This solution violates DRY in a couple places because of #@%($
type erasure, but in the end I think it provides a much better
user experience as they don't have to much with generics and play
the "what type does this return" game. by returning concrete types
for everything.

Basic ITests work but this more tests are coming. WIP commit.