Moves the query info into the future.
I went back and forth on this when I originally did it and after
starting to write docs and play with the API i decided I went the
wrong way.

This moves the "Query Info" into the RiakFuture itself. By doing so
it eliminates the need to duplicate the query info in both the
responses and the FailureInfo class I'd created. Now the core
operations simply override getQueryInfo() and return it rather than
having to include it in their respective responses. The new
netty-like RiakFuture plays fine with operations that don't
return anything since you can simply check isSuccess() after
calling await().