Fixes a number of bugs, adds ITests.
This covers fixing a number of bugs and also working around several
issues I then found with Riak. I'll be opening issues against
Riak to address those; namely a query with keyfilers and a type is
broken, and using any named function in a type other than default is
broken.

In addition I revisited how the core MR operation worked as how it
was returning things wasn't that useful.

I've done through BucketMapReduce and BucketKeyMapReduce. The remaining
two sill need the asme treatment and testing.