Puts vclock back in RiakObject.
This is another "Went the wrong way" once you start playing with the
API.

By having the vclock in the RiakObject it becomes as transparent as
possible to the end user; Get something from Riak, modify it, store
it back. Never have to muck with the vector clock.

I've modified UpdateValue to get the vclock from the fetch response
then re-inject it into the object after conflict resolution and
Update. This means the user doesn't have to deal with the vector
clock; it's taken care of for them.