Fixes problems with MR tests
There were a couple problems with the MR integration tests caused
by siblings/tombstones on the non-default bucket type.
This is now corrected by setting allow_mult to false for these
tests.

There is still one MR test that fails and it is legitimate. Using the
4-arity input (bucket, key, keydata, type) with the `default` bucket
type is still broken in Riak. You must use the 3-arity version which
does not specify a type.