Fixes RiakFuture get() implementations
The two get() methods in RiakFuture and its implementations now
conform to the behavior of core Java Future.

1) When calling either version of get() you will receive an ExecutionException
if the computation fails for any reason.
2) When calling get(timeout, timeunit) you will receive a TimeoutException
if the timer expires before the computation completes.

This means you will now either receive a response or an exception when using
get() which is the desired behavior.